melhorando main

+++ src/pages/formulario2/index.js
@@ -334,11 +334,11 @@ export default class Main extends Component {
                               <h2>{`O Total de gastos necessários é: ${
 
                         
-                                   parseInt(this.state.valor1) +
-                                   parseInt(this.state.valor2) +
-                                   parseInt(this.state.valor3) +
-                                   parseInt(this.state.valor4) +
-                                   parseInt(this.state.valor5)
+                                   parseFloat(this.state.valor1) +
+                                   parseFloat(this.state.valor2) +
+                                   parseFloat(this.state.valor3) +
+                                   parseFloat(this.state.valor4) +
+                                   parseFloat(this.state.valor5)
 
                                    }`} </h2>
